requirements.txt
carla~=0.9.13
psutil~=7.1.0
pandas~=2.0.3
matplotlib~=3.7.5
numpy~=1.24.4
scipy~=1.10.1
seaborn~=0.13.2
umap-learn~=0.5.7
kneed~=0.8.5
scikit-learn~=1.3.2
shapely~=2.0.7
pillow~=10.4.0
tqdm~=4.67.1